pyproject.toml
[project]
name = "image-to-mc"
version = "0.1.0"
description = ""
authors = [
    {name = "Naohiro2g",email = "587586+Naohiro2g@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
